backend-python/requirements.txt
# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
# sqlite3 是Python内置模块，无需安装

# 认证和安全
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# HTTP客户端和API
httpx==0.25.2
requests==2.31.0

# AI和文件处理
openai==1.3.7
PyPDF2==3.0.1
markdown==3.5.1

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1

# 其他工具
python-dotenv==1.0.0
cachetools==5.3.2